Remove printer.rs
